requirements.txt
Django>=2.2.13,<3
django-s3sign==0.1.5
pytz==2021.1
httplib2==0.19.1
feedparser==6.0.0
Markdown==3.1.1 # pyup: < 3
psycopg2==2.8.6
Pillow==8.1.1
versiontools==1.9.1
statsd==3.3.0
pep8==1.7.1
pyflakes==2.3.1
mccabe==0.6.1
configparser==5.0.2
pycodestyle==2.7.0
importlib-metadata==2.1.1  # for flake8
typing_extensions==3.10.0.0
zipp==3.1.0  # for flake8
flake8==3.9.2
tornado==6.0.3
beautifulsoup4==4.9.3
cssselect==1.1.0
lxml==4.6.3
fuzzywuzzy==0.18.0
sure==1.4.11
traitlets==4.3.3  # ipython
ipython_genutils==0.2.0  # ipython
decorator==5.0.6  # ipython
ptyprocess==0.7.0  # ipython
pexpect==4.8.0  # ipython
pathlib2==2.3.5  # ipython
pickleshare==0.7.5  # ipython
simplegeneric==0.8.1  # ipython
path.py==12.0.1  # ipython
backports.shutil_get_terminal_size==1.0.0  # ipython
wcwidth==0.2.5  # ipython
prompt_toolkit==2.0.10  # ipython
pygments==2.9.0  # ipython
scandir==1.10.0  # ipython
ipython==7.8.0
coverage==5.5
pyasn1==0.4.8
cryptography<3.3.2  # pyOpenSSL
pyOpenSSL==20.0.1
ndg-httpsclient==0.5.1
urllib3==1.26.5  # requests
chardet==4.0.0  # requests
idna==2.10  # requests
certifi==2021.5.30  # requests
requests==2.25.1
astroid==2.3.2
wrapt==1.12.1
lazy-object-proxy==1.6.0
pylint==2.4.3
six==1.16.0
ipaddress==1.0.23  # faker
faker==3.0.1  # factory_boy
factory_boy==2.12.0
ldap3==2.9
sqlparse==0.3.1
s3transfer==0.4.2
jmespath==0.10.0


botocore>=1.13.0,<1.21.0
boto3>=1.9.198,<1.18.0

python-dateutil==2.8.1

djangowind==1.2.3
django-indexer==0.3.0
django-templatetag-sugar==1.0
django-paging==0.2.5
django-annoying==0.10.6
django-appconf==1.0.4
rcssmin==1.0.6  # django-compressor
rjsmin==1.1.0  # django-compressor
django-compressor==2.4.1
django-statsd-mozilla==0.4.0
django-bootstrap4==3.0.1
django-debug-toolbar==2.2.1
django-waffle==2.2.0
django-jenkins==0.110.0
django-smoketest==1.1.2
sentry-sdk==1.1.0
django-extensions==3.1.3

django-stagingcontext==0.1.0
django-ga-context==0.1.0
django-impersonate==1.7.3
django-markwhat==1.6.2
gunicorn==19.10.0
django-infranil==1.1.0
django-flatblocks==1.0.0
django-storages==1.11.1

djangorestframework==3.12.4
rest-framework-generic-relations==2.0.0

ccnmtlsettings==1.9.0

django-reversion==3.0.5
text_unidecode==1.3

django-taggit==1.1.0

django-cacheds3storage==0.2.2

nameparser==1.0.6
oauth2==1.9.0.post1
oauthlib==3.1.1
httplib2==0.19.1
pylti==0.7.0
django-lti-provider==0.4.0

pyparsing==2.4.7
edtf==4.0.1 # pyup: <4.0.0

pbr==5.6.0 # bandit
pyyaml==5.4.1 # bandit
stevedore==3.3.0  # bandit
bandit==1.7.0

django-smtp-ssl==1.0
django-contact-us==0.4.4

entrypoints==0.3

django-courseaffils==2.2.5

mypy==0.740
mypy-extensions==0.4.3
typed-ast==1.4.3
django-stubs==1.2.0

django-widget-tweaks==1.4.8
confusable_homoglyphs==3.2.0
django-registration==3.1.2

distro==1.5.0
